First pass of MSVC warning fixes

+++ modules/tracker/rbt/include/visp3/rbt/vpRBSilhouetteCCDTracker.h
@@ -267,8 +267,8 @@ public:
    *
    * This value is between 0 and 1.
    */
-  float getMinimumMaskConfidence() const { return m_minMaskConfidence; }
-  void setMinimumMaskConfidence(float confidence)
+  double getMinimumMaskConfidence() const { return m_minMaskConfidence; }
+  void setMinimumMaskConfidence(double confidence)
   {
 
     m_minMaskConfidence = confidence;
